+++ managed_components/espressif__knob/CHANGELOG.md
@@ -1,5 +1,34 @@
 # ChangeLog
 
+## v0.1.5 - 2024-7-3
+
+###  Enhancements:
+* Support power save mode
+
+## v0.1.4 - 2023-11-23
+
+* Fix possible cmake_utilities dependency issue
+
+## v0.1.3 - 2023-6-2
+
+###  Enhancements:
+
+* Add power on knob position detection to avoid logical inversion caused by knob position
+* Change test to test_apps project
+
+## v0.1.2 - 2023-3-9
+
+### Enhancements:
+
+* Use cu_pkg_define_version to define the version of this component.
+
+## v0.1.1 - 2023-1-18
+
+### Bug Fixes:
+
+* Knob:
+  * Fix callback return usr_data root pointer, the usr_data of the relevant callback will now be returned.
+
 ## v0.1.0 - 2023-1-5
 
 ### Enhancements: